update(redirect): update for review and handling open error
