+++ srcs/exec/expand/expand_for_redirect.c
@@ -8,8 +8,6 @@
 
 static bool	is_ambiguous_redirect(t_redirect *redirect)
 {
-	if (redirect->kind == TOKEN_KIND_REDIRECT_HEREDOC)
-		return (false);
 	return (redirect->tokens->size != 1);
 }
 
@@ -19,14 +17,16 @@ static t_result	expand_for_filename_each(t_redirect *redirect, \
 	char		*original_token;
 	t_result	result;
 
+	result = SUCCESS;
 	original_token = x_ft_strdup(get_head_token_str(redirect->tokens));
-	expand_processing(&redirect->tokens, context);
+	if (expand_processing(&redirect->tokens, context) == PROCESS_ERROR)
+		return (PROCESS_ERROR);
 	split_expand_word(&redirect->tokens);
-	result = is_ambiguous_redirect(redirect);
-	if (result == FAILURE)
+	if (is_ambiguous_redirect(redirect))
 	{
 		context->status = STATUS_REDIRECT_FAILURE;
 		puterr_cmd_msg(original_token, ERROR_MSG_AMBIGUOUS);
+		result = FAILURE;
 	}
 	ft_free(&original_token);
 	return (result);
@@ -42,6 +42,7 @@ t_result	expand_for_filename(t_ast *self_node, t_context *context)
 {
 	t_deque_node	*node;
 	t_redirect		*redirect;
+	t_result		result;
 
 	node = self_node->redirect_list->node;
 	while (node)
@@ -52,8 +53,9 @@ t_result	expand_for_filename(t_ast *self_node, t_context *context)
 			node = node->next;
 			continue ;
 		}
-		if (expand_for_filename_each(redirect, context) == FAILURE)
-			return (FAILURE);
+		result = expand_for_filename_each(redirect, context);
+		if (result == FAILURE || result == PROCESS_ERROR)
+			return (result);
 		node = node->next;
 	}
 	return (SUCCESS);
